+++ spec/acceptance/xml_parser_spec.rb
@@ -1074,5 +1074,33 @@ module Resync
         expect(md.until_time).to be_time(expected_untils[i])
       end
     end
+
+    it 'parses a resource dump index' do
+      data = File.read('spec/data/resourcedump-index.xml')
+      sitemapindex = XMLParser.parse(data)
+      expect(sitemapindex).to be_a(ResourceDumpIndex)
+
+      links = sitemapindex.links
+      expect(links.size).to eq(1)
+      ln0 = links[0]
+      expect(ln0.rel).to eq('up')
+      expect(ln0.uri).to eq(URI('http://example.com/dataset1/capabilitylist.xml'))
+
+      md = sitemapindex.metadata
+      expect(md.capability).to eq('resourcedump')
+      expect(md.at_time).to be_time(Time.utc(2013, 1, 3, 9))
+      expect(md.completed_time).to be_time(Time.utc(2013, 1, 3, 9, 10))
+
+      sitemaps = sitemapindex.resources
+      expect(sitemaps.size).to eq(3)
+
+      expected_times = [Time.utc(2013, 1, 3, 9), Time.utc(2013, 1, 3, 9, 3), Time.utc(2013, 1, 3, 9, 7)]
+      (0..2).each do |i|
+        sitemap = sitemaps[i]
+        expect(sitemap.uri).to eq(URI("http://example.com/resourcedump#{i + 1}.xml"))
+        md = sitemap.metadata
+        expect(md.at_time).to be_time(expected_times[i])
+      end
+    end
   end
 end