+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## 0.1.1
 
 - Add support for [Change Dump Indices](http://www.openarchives.org/rs/1.0/resourcesync#ChangeDumpIndex)
+- Add support for [Resource Dump Indices](http://www.openarchives.org/rs/1.0/resourcesync#ResourceDumpIndex)
 
 ## 0.1.0